Issue #766 changed field for buffer size in ref from int to long

+++ bouncer/src/repo/core/handler/fileservice/repo_data_ref.cpp
@@ -32,6 +32,6 @@ DataRef DataRef::deserialise(const repo::core::model::RepoBSON &serialisedObj) {
 	return DataRef(
 		serialisedObj.getStringField(REPO_LABEL_BINARY_FILENAME),
 		serialisedObj.getIntField(REPO_LABEL_BINARY_START),
-		serialisedObj.getIntField(REPO_LABEL_BINARY_SIZE)
+		serialisedObj.getLongField(REPO_LABEL_BINARY_SIZE)
 	);
 }
